Add Flextype CMS
Flextype is an open-source self-hosted Data-First Headless CMS & API. It is was founded in March 2018 as lightweight alternative to other heavy and outdated CMS. Many people use complex solutions for simple pages, unnecessarily. Building this content management system, we focused on simplicity - even novice webmaster adapt his template and writes his own plugin. To achieve this, we implemented a simple but powerful API's.